Ditch the original erl_to_html
Instead of threading the file through all the call in erl_to_html.erl
I'm going to have all the calls build up psuedo-HTML then parse and
write it all at once.

The reason for the psuedo-HTML is that the calls and macros to make the
HTML were ugly and hard to read. The parse_atoms.erl file is where I was
experimenting with having symbol functions (e.g. '('() -> ... end) where
I turn a symbol into an atom and use it as a function name to generate
HTML. That was too clever so I switched to just keeping the symbol as an
atom and then parsing them out at the end. Anything that is an atom when
the psuedo-HTML is completed is fed to parse_symbol/1 which simply turns
those symbols into spans with class names.

Again, I could have had two competing strategies with branches but I'm
so early in this project that I don't need git for merging different
branches. I suppose there's an argument for having a topic branch right
off the start but ... I'm not going to bother. Maybe next project.