DRY. Fix extra period. Comments. Examples.
Use the module function.

There was an extra period after the function clause.

Add some comments before the file attribute and eof additions. This is
so that we can include that but show that they're inserted into the
code.

Add more example test modules that I'm using to see how the parse
transform forms are coming out. These examples will help me generate
HTML that I'll use in the test cases.