Add how-to-use, remove TMI
Add instructions on how to use this tool

Remove unnecessary "Why" reasoning. I don't need to highlight that I'm a
quitter when it comes to hobby projects. It's more efficient for
learning if I quit as soon as I've figured out what I was trying to
learn, but I don't want to bother explainging that.