Create line spans, fix < and >
Surround all of the spans on a line with a span so that the line can be
turned into a block with CSS. This also adds blank lines.

Replace < and > with &lt; and &gt;

Replace test_example_function.html with what is now generated.

Add html header and footer files