Add CSS and font
I'm trying to make the output look like my MacVim setup with the
solarized dark theme. Menlo is the font I use and I downloaded it from
http://www.cufonfonts.com/en/font/13494/menlo. colors.html shows all
the colors used in Solarized, which can be found at:
https://github.com/altercation/solarized/blob/master/vim-colors-solarized/colors/solarized.vim

I've updated the header file to include the CSS and the CSS points to
the font file. I've also fixed up the header and footer.