Remove comments, add run command, ...
The run command will run the erl_to_html:write_html command with
erl_to_html.erl as the file and output everything to the file "out".
This makes it easier to see the output, including the raw parse
transform forms.

Removed the old erl_to_html_2.erl file which I was using to figure out
what was going wrong with erl_to_html.erl when it wouldn't compile. The
"return_errors" option makes this unnecessary.