WIP: converting to tuples, cont ...
The PNG writer will operate on a stream of tuples rather than a tree
of HTML elements like a browser. There is no state carried from one
tuple to another. If an element to be rendered contains a different
element with a different style then three elements will need to be
rendered: the outer before element, the inner element and the outer
after element.

e.g.

[{1, <<"foo start">>, ?STYLE_ONE},
 {1, <<"bar">>, ?STYLE_TWO},
 {1, <<"foo end">>, ?STYLE_ONE}]

as opposed to

<div style=one>
  foo start
  <span style=two>
    bar
  </span>
  foo end
</div>