Colours as tuples of RGB instead of #RRGGBB
Converting Erlang to tuples of {LineNo, BinText, {R, G, B}} is for
erl_to_png to render the binary text and convert those the rendered
letters with the specified colour. The erl_to_png project takes pixels
that have red, green and blue values as opposed to a hex number.

This file has all the colours of the various Erlang syntax elements in
tuple form as opposed to hexadecimal.