Specify include paths, write & read tuples
In order to compile erlang files with include directives I need to
specify where to find the include files with the {i, Path} compile
option. The caller can pass in a list of paths that will be added to the
compile arguments.

Write the tuples out to a file within the parse_transform function and
then read them back in the file:consult so we can pass them back to the
caller. I was getting errors when I tried to call my own module from
within parse_transform and this worked. Good enough for now.