Sort tuples by line, retain intra-line ordering
The tuples are out of order and I don't really know why. Instead of
finding the root cause I'm sorting them before I return them. I'm
keeping the existing order of tuples when they are on the same line, so
if we have {1, X, _}, {1, Y, _}, {2, _, _}, {1, Z, _} then the order will
be {1, X, _}, {1, Y, _}, {1, Z, _}, {2, _, _}. That is, the tuples with
the same line number do not change order with respect to each other.