Hard-code in some spacing
I'm not entirely sure how I'm going to handle spacing so I'm going to
try out some manual spacing and see if I feel like that's enough.

Another option is to find all the spacing in the original file and put
it back into the output. This is similar to how comments and indents
work but I think it will be trickier: comments and indents are on the
extreme ends of the line but the spaces are interspersed throughout the
code. Not all code is unchanged when it goes through parse transform.