Re-number out-of-order lines
I suspect that I was assuming that all forms are contained on one
line and then adding a appendix to the form with the same line number.
If the forms contained in the parent form stretch over more lines then
the appendix ends up being out of order.

Also fix the binary type colour.