Space macro, fix 'of' colour, 'end' on new line
Created a macro for adding a black coloured space (which I think
erl_to_png will just turn into a transparent letter).

Fix the 'of' colour

Fix the '++' colour

Remove commented out line(Line) calls

Try to put case 'end' keywords on the next line by finding out what the
last line of the case statement is and then setting the line number of
the 'end' to be max + 1.